lerobot_teleoperator_pika/pyproject.toml
[project]
name = "lerobot_teleoperator_pika"
version = "0.1.0"
description = "LeRobot Teleoperator plugin for AgileX PIKA Sense"
readme = "README.md"
authors = [
    { name = "richardSHkim", email = "shkim@superb-ai.com" }
]
requires-python = ">=3.10"
dependencies = ["lerobot==0.4.3", "numpy>=1.24", "agx-pypika"]

[project.scripts]
lerobot-teleoperator-pika = "lerobot_teleoperator_pika:main"

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"
